reorganizing call page
