+++ src/components/Call/index.tsx
@@ -1,0 +1,3 @@
+import Call from "./Call";
+
+export default Call;
