+++ src/components/Call/useUserMedia.ts
@@ -1,0 +1,31 @@
+import { useState, useEffect } from "react";
+
+export function useUserMedia(requestedMedia: MediaStreamConstraints) {
+  const [mediaStream, setMediaStream] = useState<MediaStream>();
+
+  useEffect(() => {
+    async function enableStream() {
+      try {
+        const stream = await navigator.mediaDevices.getUserMedia(
+          requestedMedia
+        );
+        setMediaStream(stream);
+      } catch (err) {
+        // Removed for brevity
+        console.log(err);
+      }
+    }
+
+    if (!mediaStream) {
+      enableStream();
+    } else {
+      return function cleanup() {
+        mediaStream.getTracks().forEach((track) => {
+          track.stop();
+        });
+      };
+    }
+  }, [mediaStream, requestedMedia]);
+
+  return mediaStream;
+}
