+++ src/types/Session.ts
@@ -1,1 +1,7 @@
 export type Language = "en" | "es";
+
+export interface AuthInfo {
+  token: string;
+  id: number;
+  type: "inmate";
+}
